fix(telemetry): split Unix-only syscall tests to platform-specific file (#532)
Move TestLogTrace_WriteError, TestLogTrace_CloseError, and
TestOpenLogFileForAppend_MkdirAllSucceedsButRetryFails from
infra_telemetry_extended_test.go into a new
infra_telemetry_extended_unix_test.go with //go:build !windows,
following the platform_dirs_*_test.go precedent from ADR-020.

Remove the syscall import from the main test file so golangci-lint
typechecking passes on Windows.
